Change the 'addProduct' and 'deleteProduct' methods

+++ src/services/wishList.service.js
@@ -13,18 +13,30 @@ class wishListService {
 	}
 
 	static addProduct = async (wishListId, productId, SKU) => {
-		const isNewProduct = await WishList_Product.findOne({where: {wishListId, productId, SKU}});
+
+		const isNewProductPromise =  WishList_Product.findOne({where: {wishListId, productId, SKU}});
+		const wishListPromise = WishList.findByPk(wishListId);
+
+		const [isNewProduct, wishList] = await Promise.all([isNewProductPromise, wishListPromise]);
 
 		if(isNewProduct !== null) throw boom.conflict('the product had already been added');
-		const wishListProduct = await WishList_Product.create({wishListId, productId, SKU});
-		if(!(wishListProduct instanceof WishList_Product)) throw boom.badImplementation('Unexpected error');
+		if(!wishList) throw boom.notFound('Wish List not found');
+
+		const createResult = await wishList.createItem({ productId, SKU });
+		if(!(createResult instanceof WishList_Product)) throw boom.badImplementation('Unexpected error');
 		return await this.findOne(wishListId);
 	}
 
 	static deleteProduct = async (wishListId, ref) => {
+
+		const wishList = await WishList.findByPk(wishListId);
+		if(!wishList) throw boom.notFound('WishList not found');
+
 		const res = await WishList_Product.destroy({where: {ref}});
+
 		if(res === null) throw boom.badImplementation('Unexpected error');
-		if(res === 0) throw boom.badRequest("The ref is not valid");
+		if(res === 0) throw boom.notFound('Item not found');
+
 		return await this.findOne(wishListId);
 	}
